fix: info screen for both routines and workouts

+++ app/src/main/java/org/librefit/ui/screens/workout/beforeSaving/BeforeSavingScreenViewModel.kt
@@ -52,7 +52,7 @@ class BeforeSavingScreenViewModel : ViewModel() {
             it.sets.sumOf { set ->
                 if (it.setMode == SetMode.WEIGHT && set.completed) {
                     set.weight.toDouble() * set.reps
-                } else 0.toDouble()
+                } else 0.0
             }
         }.toString()
     }
